EMERGEMENCY fixs about cheat mode and item
Now all the cheat stats are contained in the getters of stats (in Hero
overriding the Living_Creatures getters)

Allowing back to normal stats.

Item's statistics addition works (on normal stats) even if you're in
cheat mode.

Changing addStats methods (for item addiction) : simple getters to
super.getters (to prevent the set(BROKEN_DAMAGE+item.getDamage() when
the player is on cheat mode)